join egnaddr and egnkey into a single crate
