+++ crates/eigen-cli/Cargo.toml
@@ -11,10 +11,24 @@ alloy-json-rpc.workspace = true
 alloy-primitives.workspace = true
 alloy-provider.workspace = true
 alloy-transport.workspace = true
+ark-ff.workspace = true
+ark-serialize = "0.4.2"
 clap.workspace = true
+eigen-crypto-bls.workspace = true
 eigen-testing-utils.workspace = true
 eigen-utils.workspace = true
+eth-keystore.workspace = true
+hex.workspace = true
+k256.workspace = true
+rand = "0.8"
+rand_core.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
+uuid = { version = "1.10.0", features = ["v4"] }
+
+[dev-dependencies]
+eigen-testing-utils.workspace = true
+rstest = "0.22.0"
+tempfile = "3"
