--- crates/eigen-cli/egnkey/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "egnkey"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-repository.workspace = true
-
-[dependencies]
-alloy-contract.workspace = true
-alloy-json-rpc.workspace = true
-alloy-primitives.workspace = true
-alloy-provider.workspace = true
-alloy-transport.workspace = true
-ark-ff.workspace = true
-ark-serialize = "0.4.2"
-clap.workspace = true
-eigen-crypto-bls.workspace = true
-eigen-utils.workspace = true
-eth-keystore.workspace = true
-hex.workspace = true
-k256.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-thiserror.workspace = true
-tokio.workspace = true
-rand = "0.8"
-rand_core.workspace = true
-uuid = { version = "1.10.0", features = ["v4"] }
-
-[dev-dependencies]
-tempfile = "3"
-rstest = "0.22.0"
-eigen-testing-utils.workspace = true
