+++ crates/eigen-cli/src/convert.rs
