+++ crates/eigen-cli/src/generate.rs
