+++ crates/eigen-cli/src/main.rs
@@ -1,23 +1,7 @@
 use clap::Parser;
-use eigen_cli::{
-    args::{Args, Commands},
-    eigen_address::ContractAddresses,
-};
+use eigen_cli::{args::Args, execute_command};
 
-#[tokio::main]
-async fn main() {
+fn main() {
     let args = Args::parse();
-    match args.command {
-        Commands::GetAddresses {
-            service_manager,
-            registry_coordinator,
-            rpc_url,
-        } => {
-            let addresses =
-                ContractAddresses::get_addresses(service_manager, registry_coordinator, rpc_url)
-                    .await
-                    .unwrap();
-            println!("{}", serde_json::to_string_pretty(&addresses).unwrap());
-        }
-    }
+    execute_command(args).unwrap();
 }
